Configuração concluída do servidor e cliente

+++ src/JogoDaVelha/Common/SocketMessage.cpp
@@ -58,6 +58,6 @@ void SocketMessage::createBuffer(char * &buffer, size_t * size)
 
 	int offset = (char*)(codePointer + 1) - (char*)codePointer;
 
-	for (int i = 0; i < _size; ++i)
+	for (size_t i = 0; i < _size; ++i)
 		buffer[i + offset] = _content[i];
 }
